+++ lib/nonogram_dart.dart
@@ -1,6 +1,7 @@
 library nonogram_dart;
 
 export 'package:nonogram_dart/src/description.dart';
+export 'package:nonogram_dart/src/generate.dart';
 export 'package:nonogram_dart/src/nonogram.dart';
 export 'package:nonogram_dart/src/solver/line_solver.dart';
 export 'package:nonogram_dart/src/solver/solver.dart';