+++ test/generate_test.dart
@@ -1,12 +1,18 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:nonogram_dart/src/generate.dart';
+import 'package:nonogram_dart/nonogram_dart.dart';
 
 void main() {
-  group('Generator', () {
-    test('monochrome', () {
-      final puzzle = Generator.monochrome(5, 5);
-      expect(puzzle.columns.length, 5);
-      expect(puzzle.rows.length, 5);
+  const b = Colors.black;
+  const w = Colors.white;
+  group('GridGenerator', () {
+    test('apply', () {
+      final gen = GridGenerator(width: 10, height: 10, colors: {b, w});
+      gen.apply([RandomizeAll()]);
+      final grid = gen.getGrid();
+
+      expect(grid.height, 10);
+      expect(grid.width, 10);
+      expect(grid.toNonogram().colors, {b});
     });
   });
 }